MOSDAC-SIMPLIFIED/requirements-minimal.txt
streamlit
langchain
langchain-community
langchain-google-genai
pypdf2
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
streamlit-chat
streamlit-option-menu
plotly
pandas
numpy
